requirements.txt
django
#uwsgi
#mysql-python
django-storages
emtex_common_utils
#tagging
django-tagging
django-extensions
django-mptt
ipython==5.0
boto==2.47.0
apscheduler
requests
djangorestframework
django-cors-headers